feat: add controllers for AnalyticDB and SimpleVector management
